==> scripts/03_Masking.Rmd <==
---
title: 'Masking'
author: 'brittth'
date: '5 May 2020'
output: word_document
---
```{r setup, include = FALSE}
knitr::opts_chunk$set(eval = FALSE)
```

# SETTINGS (Adjust them!)
```{r}
# val.id <- 'A'
# custom.crop <- c(474300,475740,5245000,5248380) #xmin,xmax,ymin(new),ymax

# val.id <- 'B'

# val.id <- 'C'

# val.id <- 'D'

# val.id <- 'E'

# val.id <- 'F'

# val.id <- 'G'

# val.id <- 'H'

# set working directory
setwd('C:/Users/brittth/Documents/Sentinel_Validation/')
```

# load packages
```{r}
# install.packages('raster')
# install.packages('ggplot2')
# install.packages('gridExtra')
# install.packages('RStoolbox')
# install.packages('grid')
# install.packages('stringr')
library(raster)
library(ggplot2)
library(gridExtra)
library(RStoolbox)
library(grid)
library(stringr)
```

# load data
```{r}
# load meta data
meta <- read.csv('Metadata.csv')

# subset meta to relevant dataset by Validation ID
meta <- meta[meta$Validation.ID %in% val.id,]

# adjust val.ids
val.id.file <- chartr(' ', '_', val.id)
val.ids <- c(paste0(val.id.file,'_Land'),paste0(val.id.file,'_Water'))

# set input folder path for spectrally and spatially resampled reference, sentinel data and scl data
file.list <- list.files(val.ids[1], pattern= 'spatres',full.names = TRUE)
  # val.id not important here as reference and sentinel imagery is identical for both validation ids

# set metainfo
  # Sentinel-2
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){
  datanames <- c('reference10', 'reference20', 'reference60','sentinel10', 'sentinel20', 'sentinel60')
  bandnames <- list(c('Band2','Band3','Band4','Band8'), # ref 10m
                  c('Band5','Band6','Band7','Band8A','Band11','Band12'),# ref 20m
                  c('Band1','Band9'),# ref 60m,
                  c('Band2','Band3','Band4','Band8'), # sen 10m
                  c('Band5','Band6','Band7','Band8A','Band11','Band12'),# sen 20m
                  c('Band1','Band9'))# sen 60m
  # Sentinel-3
}else{
  datanames <- c('reference300', 'sentinel300')
  bandnames <- list(c('Band1','Band2','Band3','Band4','Band5','Band6','Band7','Band8','Band9','Band10',
                      'Band11','Band12','Band16','Band17','Band18','Band21'),# ref
                    c('Band1','Band2','Band3','Band4','Band5','Band6','Band7','Band8','Band9','Band10',
                      'Band11','Band12','Band16','Band17','Band18','Band21'))# sen
}


# imagery
im.list <- list()

for (i in 1:length(file.list)){
# i=5
  # load spectrally and spatially resampled reference and sentinel imagery
  im <- brick(file.list[[i]])
  
  # update bandnames
  names(im) <- bandnames[[i]]
  # plot(im)
  
  # store in image list
  im.list[[i]] <- im
}

if(meta$Sentinel == 'Sentinel-3A' | meta$Sentinel == 'Sentinel-3B'){
# include scaling factor to make datasets comparable (actual numbers are not important)
im.list[[2]] <- im.list[[2]]*10000
}

# SCL

# check if alternative mask is required
input <- read.csv('scripts/input/input.csv')
input <- input[input$Validation.ID %in% val.id,]
if(nchar(as.character(input$Sentinel.Mask.Path))>0){
  scl.alt <- brick(as.character(input$Sentinel.Mask.Path))
}

  # Sentinel-2
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){

  # load scene classification layer at 20m resolution (doesn't exist for 10m)
  if(exists('scl.alt')){
    scl <- scl.alt
  }else{
    scl <- brick(paste0(meta$Sentinel.Scene.Path,'R20m/', meta$Sentinel.Scene.Short, '_SCL_20m.jp2'))
  }

  # crop Sentinel-2 SCL image toAOI extent
  scl <- crop(scl,extent(im.list[[4]]), snap='in') #any sentinel image works
  
  # get outline of reference/sentinel imagery
  im.outline.r <- im.list[[1]][[1]] > -Inf

  # convert to polygon
  im.outline <- rasterToPolygons(im.outline.r, dissolve=TRUE)

  # crop flag layer to reference/sentinel imagery outline
  scl <- mask(scl, im.outline)

  # Sentinel-3
}else{

  # load flag layer
  if(exists('scl.alt')){
    syn <- scl.alt}
  else{
    syn <- brick(paste0(meta$Sentinel.Scene.Path,meta$Sentinel.Scene.Short,meta$Sentinel.Scene.Filetype))
  }
  scl <- syn[[17]]

  # get outline of reference/sentinel imagery
  im.outline.r <- im.list[[1]] > -Inf
  
  # project raster and convert to polygon
  im.outline.r <- projectRaster(im.outline.r, crs= crs(scl))
  im.outline <- rasterToPolygons(im.outline.r, dissolve=TRUE)

  # crop flag layer to reference/sentinel imagery outline
  scl <- mask(scl, im.outline)
  scl <- trim(scl)
}

# plot(scl, colNA='red')
# unique(values(scl))
```

# create mask buffer to exclude outliers - 120m general buffer
```{r}  
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){ # Sentinel-2
  
  # prepare mask buffer
  buffer <- scl$layer
  buffer[buffer == 4 | buffer == 5 | buffer == 6] <- NA
  plot(buffer, colNA='red')
  
}else{ # Sentinel-3
  # prepare mask buffer
  buffer <- scl$tif_subset_bsub_UTM32Nproj.17
  buffer[buffer == (4096 + 16) | buffer == (4096 + 256 + 16) | buffer == 0] <- NA
  plot(buffer, colNA='red')
}

# create mask buffer
buffer <- buffer(buffer, width=120)
# plot(buffer, colNA='red')

# convert
buffer.poly <- rasterToPolygons(buffer, fun=NULL, n=16, na.rm=TRUE, digits=12, dissolve=TRUE)
# plot(buffer.poly, col=c('grey'))

# reproject flag layer polygon to reference/sentinel imagery projection if necessary
if(as.character(crs(buffer.poly)) != as.character(crs(im))){
  buffer.poly <- spTransform(buffer.poly,crs(im))
  buffer <- projectRaster(buffer, crs= crs(im))
}
```

# category information
```{r}
# Sentinel-2 SCL Categories:
# 0 (no data)                         MASK
# 1 (saturated or defective)          MASK
# 2 (dark feature shadow)             MASK
# 3 (cloud shadows)                   MASK
# 4 (vegetation)               LAND
# 5 (non-vegetated)            LAND
# 6 (water)                    WATER
# 7 (unclassified)                    MASK
# 8 (cloud medium probability)        MASK
# 9 (cloud high probability)          MASK
# 10 (thin cirrus)                    MASK
# 11 (snow or ice)                    MASK

# Sentinel-3 Flag Layer Catgories (values added up, here Grosseto relevant categories)
# 4372  = 4096 + 256 + 16 + 4 (SYN_success + SYN_partly_cloudy + SYN_land + SYN_SDR_OOR)      MASK
# 4368  = 4096 + 256 + 16     (SYN_success  + SYN_partly_cloudy + SYN_land)             LAND
# 4116  = 4096 + 16 + 4       (SYN_success + SYN_land + SYN_SDR_OOR)                          MASK
# 4112  = 4096 + 16           (SYN_success + SYN_land)                                  LAND             
# 17    = 16 + 1              (SYN_land + SYN_cloud)                                          MASK
# 0     (?)                                                                             WATER
```

# create mask for land and water as raster and polygon
```{r}
# get SCL
scl.land <- scl
scl.water <- scl


# Sentinel-2
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){

  # create mask for land surfaces
  scl.land[scl.land != 4 & scl.land != 5] <- NA
  
  # create mask for water surfaces
  scl.water[scl.water != 6] <- NA
  
# Sentinel-3
}else{
  
  # create mask for land surfaces
  scl.land[scl.land != (4096 + 16) & scl.land != (4096 + 256 + 16)] <- NA
  
  # create mask for water surfaces
  scl.water[scl.water != 0] <- NA
}

# unite all values to 1
scl.land <- scl.land > -Inf
scl.water <- scl.water > -Inf

# convert raster to spatial polygon (SpatialPolygonDataframe) --> avoid extent error
if(all(is.na(values(scl.land)))){
  # write note
  write.table(NULL, file = paste0(val.ids[1],'/IMAGERY_DOES_NOT_CONTAIN_LAND_PIXELS.txt'))
  cat('\nIMAGERY DOES NOT CONTAIN LAND PIXELS !')
  
}else{
  # trim away NAs
  scl.land <- trim(scl.land)
  
  # convert 
  scl.land.poly <- rasterToPolygons(scl.land, fun=NULL, n=16, na.rm=TRUE, digits=12, dissolve=TRUE)
  plot(scl.land.poly, col=c('chocolate'))
  
  # reproject flag layer polygon to reference/sentinel imagery projection if necessary
  if(as.character(crs(scl.land.poly)) != as.character(crs(im))){
    scl.land.poly <- spTransform(scl.land.poly,crs(im))
  }
  
}

if(all(is.na(values(scl.water)))){
  # write note
  write.table(NULL, file = paste0(val.ids[2],'/IMAGERY_DOES_NOT_CONTAIN_WATER_PIXELS.txt'))
  cat('\nIMAGERY DOES NOT CONTAIN WATER PIXELS !')
  
}else{
  # trim away NAs
  scl.water <- trim(scl.water)
  
  # convert
  scl.water.poly <- rasterToPolygons(scl.water, fun=NULL, n=16, na.rm=TRUE, digits=12, dissolve=TRUE)
  plot(scl.water.poly, col=c('blue'))
  
  # reproject flag layer polygon to reference/sentinel imagery projection
  if(as.character(crs(scl.water.poly)) != as.character(crs(im))){
    scl.water.poly <- spTransform(scl.water.poly,crs(im))
  }
  
}

# reproject raster layers (has to be done after polygon creation to not distort values)
if(as.character(crs(scl)) != as.character(crs(im))){
  scl <- projectRaster(scl, crs= crs(im))
  scl.land <- projectRaster(scl.land, crs= crs(im))
  scl.water <- projectRaster(scl.water, crs= crs(im))
}
```

# apply masks
```{r}
im.land.list <- list()
im.water.list <- list()

# go through all images
# i=1
for (i in 1:length(im.list)){
  
  # get image
  im <- im.list[[i]]

  # LAND
  if(all(is.na(values(scl.land)))){ # could also simply check if poly exists
    
    # call list slot NA
    im.land <- NA
    insufficient.land.px <- TRUE

  }else{
    
    # apply SpatialPolygone version of mask (else problems when resolution differs)
    im.land <- mask(im, scl.land.poly)
    plot(im.land)
    
    # apply SpatialPolygone version of mask buffer
    im.land <- mask(im.land, buffer.poly,inverse=TRUE)
    plot(im.land)
    
    # custom cropping
    if(exists('custom.crop')){#so far only for Switzerland Greifensee
      crop.extent <- extent(custom.crop)
      im.land <- crop(im.land, crop.extent) 
    }
    
    # check if enough land pixels remain after masking
    if(nrow(na.omit(as.data.frame(im.land))) < 30){
      insufficient.land.px <- TRUE
      write.table(NULL, file = paste0(val.ids[1],'/IMAGERY_DOES_NOT_CONTAIN_ENOUGH_LAND_PIXELS.txt'))
      cat('\nIMAGERY DOES NOT CONTAIN ENOUGH LAND PIXELS !')
      im.land <- NA
      
    }else{
      # trim away NAs
      im.land <- trim(im.land)
    
      # write masked imagery to disc
      writeRaster(im.land, paste0(val.ids[1],'/',val.ids[1],'_',datanames[i],'.tif'),overwrite=TRUE)
      cat(paste0('\nValidation ID: ',paste0(val.id,'  Land'),'\nImage ',datanames[i],' successfully masked for Land!\n'))
    }
  }
  
  
  # WATER
  if(all(is.na(values(scl.water)))){
    
    # call list slot NA
    im.water <- NA
    insufficient.water.px <- TRUE
  
  }else{
    
    # apply SpatialPolygone version of mask (else problems when resolution differs)
    im.water <- mask(im,scl.water.poly)
    # plot(im.water)
    
    # apply SpatialPolygone version of mask buffer
    im.water <- mask(im.water, buffer.poly,inverse=TRUE)
    # plot(im.water)
    
    # check if enough land pixels remain after masking
    if(nrow(na.omit(as.data.frame(im.water))) < 30){
      insufficient.water.px <- TRUE
      write.table(NULL, file = paste0(val.ids[2],'/IMAGERY_DOES_NOT_CONTAIN_ENOUGH_WATER_PIXELS.txt'))
      cat('\nIMAGERY DOES NOT CONTAIN ENOUGH WATER PIXELS !')
      im.water <- NA
      
    }else{
      # trim away NAs
      im.water <- trim(im.water)
      
      # write masked imagery to disc
      writeRaster(im.water, paste0(val.ids[2],'/',val.ids[2],'_',datanames[i],'.tif'),overwrite=TRUE)
      cat('\nValidation ID: ',paste0(val.id,'  Water'),'\nImage ',datanames[i],' successfully masked for Water!\n')
    }
    
     
  }
  
  
  # add to list
  im.land.list[[i]] <- im.land
  im.water.list[[i]] <- im.water
  
}

# create map outline for custom crop
if(exists('custom.crop')){#so far only for Switzerland Greifensee
  custom.crop.outline <- as(crop.extent, 'SpatialPolygons')
  # plot(custom.crop.outline)
}
# plot(scl.land.poly, col= c('chocolate','chocolate'))
# plot(im.land.list[[1]][[1]], colNA='red')
```

# common map elements
```{r}
box <- as(extent(scl), 'SpatialPolygons')
# box.scl.path <- geom_path(aes(x = long, y = lat), data = box.scl, size=0.1, col='white') # relic?
box.path <- geom_path(aes(x = long, y = lat), data = box, size=0.1, col='white')

commontheme <- theme_void()+theme(plot.title = element_text(size=12, face='bold',hjust = 0.5))
not.available <- textGrob('X',gp = gpar(fontsize=40, fontface ='bold', col='red'))

if(exists('custom.crop')){
  cc.outline.path <- geom_path(aes(x = long, y = lat), data = custom.crop.outline, 
                               size=1, col='darkgrey', linetype='longdash')
  anno.cc <- annotate('text',col='darkgrey', label = 'custom crop', 
                      x = mean(custom.crop[1:2]), y = custom.crop[3]-100, fontface ='bold', size = 2.5)}
```

# create mask maps
```{r}
# Sentinel-2
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){

  # prepare scl dataframe for plotting
  scl.df <- as.data.frame(scl, xy=TRUE)
  colnames(scl.df)[3] <- 'layer'
  scl.df$class <- NA
  scl.df[scl.df$layer %in% 0,'class'] <- 'no data'
  scl.df[scl.df$layer %in% 1,'class'] <- 'saturated or defective'
  scl.df[scl.df$layer %in% 2,'class'] <- 'dark feature shadow'
  scl.df[scl.df$layer %in% 3,'class'] <- 'cloud shadows'
  scl.df[scl.df$layer %in% 4,'class'] <- 'vegetation (LAND MASK)' 
  scl.df[scl.df$layer %in% 5,'class'] <- 'non-vegetated (LAND MASK)'
  scl.df[scl.df$layer %in% 6,'class'] <- 'water (WATER MASK)'
  scl.df[scl.df$layer %in% 7,'class'] <- 'unclassified'
  scl.df[scl.df$layer %in% 8,'class'] <- 'cloud medium probability'
  scl.df[scl.df$layer %in% 9,'class'] <- 'cloud high probability'
  scl.df[scl.df$layer %in% 10,'class'] <- 'thin cirrus'
  scl.df[scl.df$layer %in% 11,'class'] <- 'snow or ice'
  scl.df[scl.df$layer %in% NA,'class'] <- ' '
  scl.cols <- c('no data' = 'darkred',
                'saturated or defective' = 'red', 
                'dark feature shadow' = 'black', 
                'cloud shadows' = 'darkgrey',
                'vegetation (LAND MASK)' = 'darkgreen', 
                'non-vegetated (LAND MASK)' = 'orange',
                'water (WATER MASK)' = 'blue', 
                'unclassified' = 'firebrick1',
                'cloud medium probability' = 'grey', 
                'cloud high probability' = 'lightgrey',
                'thin cirrus' = 'lightblue', 
                'snow or ice' = 'royalblue',
                ' '='white')
  
  # mask plot title
  mask.title <- 'Scene Classification Layer (SCL)'
  
  
# Sentinel-3
}else{

  # prepare scl dataframe for plotting
  scl.df <- as.data.frame(scl, xy=TRUE)
  scl.df$class <- NA
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 0,'class'] <- 'water (WATER MASK)\n'
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 17,'class'] <- 'SYN_land + \nSYN_cloud\n'
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 4112,'class'] <- 'SYN_success + \nSYN_land (LAND MASK)\n'
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 4116,'class'] <- 'SYN_success + \nSYN_land + \nSYN_SDR_OOR\n'
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 4368,'class'] <- 'SYN_success  + \nSYN_partly_cloudy + \nSYN_land (LAND MASK)\n' 
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% 4372,'class'] <- 'SYN_success + \nSYN_partly_cloudy + \nSYN_land + \nSYN_SDR_OOR\n'
  scl.df[scl.df$tif_subset_bsub_UTM32Nproj.17 %in% NA,'class'] <- ''
  scl.cols <- c('water (WATER MASK)\n' = 'blue',
                'SYN_land + \nSYN_cloud\n' = 'lightgrey',
                'SYN_success + \nSYN_land (LAND MASK)\n' = 'chocolate',
                'SYN_success + \nSYN_land + \nSYN_SDR_OOR\n' = 'firebrick1',
                'SYN_success  + \nSYN_partly_cloudy + \nSYN_land (LAND MASK)\n' = 'orange',
                'SYN_success + \nSYN_partly_cloudy + \nSYN_land + \nSYN_SDR_OOR\n' = 'red',
                ' '='white')
  
  # mask plot title
  mask.title <- 'Flag Layer'
}

# create mask maps
scl.plot <- ggplot(scl.df) +  
  geom_raster(aes(x=x, y=y, fill=class))+
  box.path+
  coord_equal()+
  commontheme +
  theme(legend.title = element_blank())+
  labs(title = mask.title)+
  scale_fill_manual(values = scl.cols)

if(all(is.na(values(scl.land)))){
  not.available.title <- textGrob('Land Mask',gp = gpar(fontface ='bold', col='black'))
  scl.land.plot <- grid.arrange(not.available, top = not.available.title)
  
}else{
  # prepare scl.land dataframe for plotting
  scl.land.df <- as.data.frame(scl.land, xy=TRUE)
  scl.land.df[!scl.land.df$layer %in% NA,'layer'] <- 'land'
  scl.land.df[scl.land.df$layer %in% NA,'layer'] <- ' '
  scl.land.cols <- c('land' = 'chocolate', ' '='white')
  
  scl.land.plot <- ggplot(scl.land.df)+ 
    geom_raster(aes(x=x, y=y, fill=layer), show.legend = FALSE)+
    box.path+
    coord_equal()+
    commontheme +
    labs(title = 'Land Mask')+
    scale_fill_manual(values = scl.land.cols)
}

if(all(is.na(values(scl.water)))){
  not.available.title <- textGrob('Water Mask',gp = gpar(fontface ='bold', col='black'))
  scl.water.plot <-   grid.arrange(not.available, top = not.available.title)
   
}else{
  # prepare scl.water dataframe for plotting
  scl.water.df <- as.data.frame(scl.water, xy=TRUE)
  scl.water.df[!scl.water.df$layer %in% NA,'layer'] <- 'water'
  scl.water.df[scl.water.df$layer %in% NA,'layer'] <- ' '
  scl.water.cols <- c('water' = 'blue', ' '='white')
  
  scl.water.plot <- ggplot(scl.water.df)+ 
    geom_raster(aes(x=x, y=y, fill=layer), show.legend = FALSE)+
    box.path+
    coord_equal()+
    commontheme +
    labs(title = 'Water Mask')+
    scale_fill_manual(values = scl.water.cols)
}

if(exists('custom.crop')){
  scl.plot <- scl.plot + cc.outline.path + anno.cc
  scl.land.plot <- scl.land.plot + cc.outline.path + anno.cc
}

# create mask overview
title <- textGrob(val.id,gp = gpar(fontsize=14, fontface ='bold'))

# arrangement
if(val.id == 'A'){ # Greifensee-specific
  mask.map.plot <- grid.arrange(scl.plot,scl.land.plot,scl.water.plot,
                                nrow=1,ncol=3, widths=c(2,1,1),top=title)
}else{
  lay.masks <- rbind(c(1,2),
                     c(1,3))
  mask.map.plot <- grid.arrange(scl.plot,scl.land.plot,scl.water.plot,
                                  nrow=2,ncol=2, widths=c(2,1),layout_matrix = lay.masks,
                                  top=title)
}
```

# create masked image maps
```{r}
# validation titles
title.land <- textGrob(paste0(val.id,'  Land'),gp = gpar(fontsize=14, fontface ='bold'))
title.water <- textGrob(paste0(val.id,'  Water'),gp = gpar(fontsize=14, fontface ='bold'))

# no stretch possible if only 1 pixel left


# Sentinel-2
if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){
  
  lay <- rbind(c(1,2,3,4),
               c(1,5,6,7))
  
  # LAND
  if(!exists('insufficient.land.px')){
  
    # reference imagery
    ref10.land.plot <- ggRGB(im.land.list[[1]],r=3, g=2, b=1, stretch='hist')+
      commontheme + box.path + labs(title = '\n10m')
    ref20.land.plot <- ggRGB(im.land.list[[2]],r=1, g=2, b=3, stretch='hist')+
      commontheme + box.path + labs(title = 'Reference Scene\n20m')
    ref60.land.plot <- ggRGB(im.land.list[[3]],r=1, g=2, b=2, stretch='hist')+
      commontheme + box.path + labs(title = '\n60m')
    
    # sentinel imagery
    sen10.land.plot <- ggRGB(im.land.list[[4]],r=3, g=2, b=1, stretch='hist')+
      commontheme + box.path + labs(title = '\n10m')
    sen20.land.plot <- ggRGB(im.land.list[[5]],r=1, g=2, b=3, stretch='hist')+
      commontheme + box.path + labs(title = 'Sentinel Scene\n20m')
    sen60.land.plot <- ggRGB(im.land.list[[6]],r=1, g=2, b=2, stretch='hist')+
      commontheme + box.path + labs(title = '\n60m')
      
    if(exists('custom.crop')){
      ref10.land.plot <- ref10.land.plot + cc.outline.path + anno.cc
      ref20.land.plot <- ref20.land.plot + cc.outline.path + anno.cc    
      ref60.land.plot <- ref60.land.plot + cc.outline.path + anno.cc    
      sen10.land.plot <- sen10.land.plot + cc.outline.path + anno.cc
      sen20.land.plot <- sen20.land.plot + cc.outline.path + anno.cc
      sen60.land.plot <- sen60.land.plot + cc.outline.path + anno.cc
    }
  
    # arrangement
    masked.map.land.plot <- grid.arrange(scl.land.plot,
                                       sen10.land.plot,sen20.land.plot,sen60.land.plot,
                                       ref10.land.plot,ref20.land.plot,ref60.land.plot,
                                       nrow=2,ncol=4, layout_matrix = lay, 
                                       heights=c(5,5), widths=c(2,1,1,1),
                                       top = title.land)
  }  
    # WATER
  if(!exists('insufficient.water.px')){
  
    # reference imagery
    ref10.water.plot <- ggRGB(im.water.list[[1]],r=3, g=2, b=1, stretch='hist')+
      commontheme + box.path + labs(title = '\n10m')
    ref20.water.plot <- ggRGB(im.water.list[[2]],r=1, g=2, b=3, stretch='hist')+
      commontheme + box.path + labs(title = 'Reference Scene\n20m')
    ref60.water.plot <- ggRGB(im.water.list[[3]],r=1, g=2, b=2, stretch='hist')+
      commontheme + box.path + labs(title = '\n60m')
    
    # sentinel imagery
    sen10.water.plot <- ggRGB(im.water.list[[4]],r=3, g=2, b=1, stretch='hist')+
      commontheme + box.path + labs(title = '\n10m')
    sen20.water.plot <- ggRGB(im.water.list[[5]],r=1, g=2, b=3, stretch='hist')+
      commontheme + box.path + labs(title = 'Sentinel Scene\n20m')
    sen60.water.plot <- ggRGB(im.water.list[[6]],r=1, g=2, b=2, stretch='hist')+
      commontheme + box.path + labs(title = '\n60m')
    
    # arrangement
    masked.map.water.plot <- grid.arrange(scl.water.plot,
                                       sen10.water.plot,sen20.water.plot,sen60.water.plot,
                                       ref10.water.plot,ref20.water.plot,ref60.water.plot,
                                       nrow=2,ncol=4, layout_matrix = lay, 
                                       heights=c(5,5), widths=c(2,1,1,1),
                                       top = title.water)
  }
# Sentinel-3
}else{
  lay <- rbind(c(1,2),
               c(1,3))
  # LAND
  if(!exists('insufficient.land.px')){
  
    # reference imagery
    ref300.land.plot <- ggRGB(im.land.list[[1]],r=8, g=6, b=4, stretch='hist')+
      commontheme + box.path + labs(title = 'Reference Scene\n300m')  
  
    # sentinel imagery
    sen300.land.plot <- ggRGB(im.land.list[[2]],r=8, g=6, b=4, stretch='hist')+
      commontheme + box.path + labs(title = 'Sentinel Scene\n300m')
      
    if(exists('custom.crop')){
      ref300.land.plot <- ref300.land.plot + cc.outline.path + anno.cc
      sen300.land.plot <- sen300.land.plot + cc.outline.path + anno.cc    
    }
  
    # arrangement
    masked.map.land.plot <- grid.arrange(scl.land.plot,
                                       sen300.land.plot,ref300.land.plot,
                                       nrow=2,ncol=2, layout_matrix = lay, 
                                       heights=c(5,5), widths=c(2,1),
                                       top = title.land)
  }  
    # WATER
  if(!exists('insufficient.water.px')){
  
    # reference imagery
    ref300.water.plot <- ggRGB(im.water.list[[1]],r=8, g=6, b=4, stretch='hist')+ 
      commontheme + box.path + labs(title = 'Reference Scene\n300m')
    
    # sentinel imagery
    sen300.water.plot <- ggRGB(im.water.list[[2]],r=8, g=6, b=4, stretch='hist')+
      commontheme + box.path + labs(title = 'Sentinel Scene\n300m')
    
    # arrangement
    masked.map.water.plot <- grid.arrange(scl.water.plot,
                                       sen300.water.plot,ref300.water.plot,
                                       nrow=2,ncol=2, layout_matrix = lay, 
                                       heights=c(5,5), widths=c(2,1),
                                       top = title.water)
  }
}


if(val.id == 'A'){
  sen.title <- textGrob('Sentinel Scene',gp = gpar(fontsize=12, fontface ='bold'))
  ref.title <- textGrob('Reference Scene',gp = gpar(fontsize=12, fontface ='bold'))
  
  # arrangement matrix
  lay <- rbind(c(1,2,2,2,3,3,3),
               c(4,5,6,7,8,9,10))
  
  # custom titles
  ref10.land.plot <- ref10.land.plot + labs(title = '10m')  
  ref20.land.plot <- ref20.land.plot + labs(title = '20m')
  ref60.land.plot <- ref60.land.plot + labs(title = '60m')
  sen10.land.plot <- sen10.land.plot + labs(title = '10m')
  sen20.land.plot <- sen20.land.plot + labs(title = '20m')
  sen60.land.plot <- sen60.land.plot + labs(title = '60m')
  
  ref10.water.plot <- ref10.water.plot + labs(title = '10m')
  ref20.water.plot <- ref20.water.plot + labs(title = '20m')
  ref60.water.plot <- ref60.water.plot + labs(title = '60m')
  sen10.water.plot <- sen10.water.plot + labs(title = '10m')
  sen20.water.plot <- sen20.water.plot + labs(title = '20m')
  sen60.water.plot <- sen60.water.plot + labs(title = '60m')
  
  
  # overview
  masked.map.land.plot <- grid.arrange(rectGrob(gp=gpar(col=NA)),sen.title,ref.title,
                                        scl.land.plot,sen10.land.plot,sen20.land.plot,sen60.land.plot,
                                        ref10.land.plot,ref20.land.plot,ref60.land.plot,
                                        nrow=2,ncol=7, layout_matrix = lay,
                                        heights=c(1,10), widths=c(2,1,1,1,1,1,1),
                                        top = title.land)
  
  masked.map.water.plot <- grid.arrange(rectGrob(gp=gpar(col=NA)),sen.title,ref.title,
                                        scl.water.plot,sen10.water.plot,sen20.water.plot,sen60.water.plot,
                                        ref10.water.plot,ref20.water.plot,ref60.water.plot,
                                        nrow=2,ncol=7, layout_matrix = lay, 
                                        heights=c(1,10), widths=c(2,1,1,1,1,1,1),
                                        top = title.water)
}
```

# write maps to disc
```{r}
for(i in 1:length(val.ids)){

  # write mask maps to disc
  if(val.id == 'A'){ # Greifensee-specific
    ggsave(file=paste0(val.ids[i],'/',val.ids[i],'_mask_maps.png'),plot=mask.map.plot, 
               width=10, height=7, dpi = 300)
  }else{
    ggsave(file=paste0(val.ids[i],'/',val.ids[i],'_mask_maps.png'),plot=mask.map.plot, 
               width=10, height=5, dpi = 300)
  }
}


# write masked image maps to disc
  # LAND
if(!exists('insufficient.land.px')){
  # writeRaster(scl.land, paste0(val.ids[1],'/',val.ids[1],'_scl_mask.tif'),overwrite=TRUE) # optional (takes long)
  
      # Sentinel-2
  if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){
    if(val.id == 'A'){ # Greifensee-specific
      ggsave(file=paste0(val.ids[1],'/',val.ids[1],'_masked_maps.png'),plot=masked.map.land.plot,
             width=10, height=7, dpi = 300)
    }else{
      ggsave(file=paste0(val.ids[1],'/',val.ids[1],'_masked_maps.png'),plot=masked.map.land.plot,
             width=10, height=5, dpi = 300)
    }
    # Sentinel-3
  }else{
    ggsave(file=paste0(val.ids[1],'/',val.ids[1],'_masked_maps.png'),plot=masked.map.land.plot,
             width=6, height=5, dpi = 300)
  }

}
  # WATER
if(!exists('insufficient.water.px')){
  # writeRaster(scl.water, paste0(val.ids[2],'/',val.ids[2],'_scl_mask.tif'),overwrite=TRUE) # optional (takes long)
      
    # Sentinel-2
  if(meta$Sentinel == 'Sentinel-2A' | meta$Sentinel == 'Sentinel-2B'){
    if(val.id == 'A'){ # Greifensee-specific
      ggsave(file=paste0(val.ids[2],'/',val.ids[2],'_masked_maps.png'),plot=masked.map.water.plot,
             width=10, height=7, dpi = 300)
    }else{
      ggsave(file=paste0(val.ids[2],'/',val.ids[2],'_masked_maps.png'),plot=masked.map.water.plot,
               width=10, height=5, dpi = 300)
  }
    # Sentinel-3
  }else{
    ggsave(file=paste0(val.ids[2],'/',val.ids[2],'_masked_maps.png'),plot=masked.map.water.plot,
             width=6, height=5, dpi = 300)    
  }

}
```


```{r}
```
